feat : dailry 검증 로직 구현

+++ backend/src/main/java/com/daily/daily/dailry/service/DailryService.java
@@ -1,11 +1,12 @@
 package com.daily.daily.dailry.service;
 
+import com.daily.daily.common.exception.UnauthorizedAccessException;
 import com.daily.daily.dailry.domain.Dailry;
 import com.daily.daily.dailry.dto.DailryDTO;
 import com.daily.daily.dailry.dto.DailryFindDTO;
 import com.daily.daily.dailry.dto.DailryUpdateDTO;
+import com.daily.daily.dailry.exception.DailryNotFoundException;
 import com.daily.daily.dailry.repository.DailryRepository;
-import com.daily.daily.dailrypage.exception.DailryPageNotFoundException;
 import com.daily.daily.member.domain.Member;
 import com.daily.daily.member.exception.MemberNotFoundException;
 import com.daily.daily.member.repository.MemberRepository;
@@ -39,11 +40,11 @@ public class DailryService {
 
     public DailryDTO update(Long memberId, Long dairlyId, DailryUpdateDTO dailryUpdateDTO) {
         Dailry dailry = dailryRepository.findById(dairlyId)
-                .orElseThrow(DailryPageNotFoundException::new);
+                .orElseThrow(DailryNotFoundException::new);
 
-//        if(!Objects.equals(memberId, dailry.getMember().getId())) {
-//            throw new UnauthorizedAccessException();
-//        }
+        if (!dailry.belongsTo(memberId)) {
+            throw new UnauthorizedAccessException();
+        }
 
         dailry.updateTitle(dailryUpdateDTO.getTitle());
 
@@ -52,11 +53,11 @@ public class DailryService {
 
     public DailryDTO find(Long memberId, Long dairlyId) {
         Dailry dailry = dailryRepository.findById(dairlyId)
-                .orElseThrow(DailryPageNotFoundException::new);
+                .orElseThrow(DailryNotFoundException::new);
 
-//        if(!Objects.equals(memberId, dailry.getMember().getId())) {
-//            throw new UnauthorizedAccessException();
-//        }
+        if (!dailry.belongsTo(memberId)) {
+            throw new UnauthorizedAccessException();
+        }
         return DailryDTO.from(dailry);
     }
 
@@ -66,10 +67,10 @@ public class DailryService {
 
     public void delete(Long memberId, Long dairlyId) {
         Dailry dailry = dailryRepository.findById(dairlyId)
-                .orElseThrow(DailryPageNotFoundException::new);
-//        if(!Objects.equals(memberId, dailry.getMember().getId())) {
-//            throw new UnauthorizedAccessException();
-//        }
+                .orElseThrow(DailryNotFoundException::new);
+        if (!dailry.belongsTo(memberId)) {
+            throw new UnauthorizedAccessException();
+        }
         dailryRepository.deleteById(dairlyId);
     }
